+++ jksoa-mq/jksoa-mq-client/src/main/kotlin/net/jkcode/jksoa/mq/producer/MqProducer.kt
@@ -1,10 +1,12 @@
 package net.jkcode.jksoa.mq
 
+import net.jkcode.jkmvc.common.getWritableFinalField
 import net.jkcode.jksoa.client.referer.Referer
 import net.jkcode.jksoa.mq.broker.service.IMqBrokerLeaderService
 import net.jkcode.jksoa.mq.broker.service.IMqBrokerService
 import net.jkcode.jksoa.mq.common.Message
 import net.jkcode.jksoa.mq.producer.IMqProducer
+import java.util.concurrent.CompletableFuture
 
 /**
  * 消息生产者
@@ -12,6 +14,11 @@ import net.jkcode.jksoa.mq.producer.IMqProducer
  * @date 2019-01-10 8:41 PM
  */
 object MqProducer : IMqProducer {
+
+    /**
+     * 消息的id属性
+     */
+    private val idProp = Message::class.java.getWritableFinalField("id")
 
     /**
      * 消息中转者的leader
@@ -26,29 +33,34 @@ object MqProducer : IMqProducer {
     /**
      * 注册主题
      * @param topic 主题
-     * @return
+     * @return false表示没有broker可分配
      */
-    public override fun registerTopic(topic: String){
-        brokerLeaderService.registerTopic(topic)
+    public override fun registerTopic(topic: String): Boolean {
+        return brokerLeaderService.registerTopic(topic)
     }
 
     /**
      * 注销topic
      *
      * @param topic
-     * @return
+     * @return false表示topic根本就没有分配过
      */
-    public override fun unregisterTopic(topic: String){
-        brokerLeaderService.unregisterTopic(topic)
+    public override fun unregisterTopic(topic: String): Boolean {
+        return brokerLeaderService.unregisterTopic(topic)
     }
 
     /**
      * 生产消息
      * @param msg 消息
+     * @return broker生成的消息id
      */
-    public override fun send(msg: Message){
+    public override fun send(msg: Message): CompletableFuture<Long> {
         // 通过中转者来分发消息
-        brokerService.putMessage(msg)
+        return brokerService.putMessage(msg).thenApply { id ->
+            // 回写broker生成的消息id
+            idProp.set(msg, id)
+            id
+        }
     }
 
 }