+++ jksoa-mq/jksoa-mq-common/src/main/kotlin/net/jkcode/jksoa/mq/producer/IMqProducer.kt
@@ -1,6 +1,7 @@
 package net.jkcode.jksoa.mq.producer
 
 import net.jkcode.jksoa.mq.common.Message
+import java.util.concurrent.CompletableFuture
 
 /**
  * 消息生产者
@@ -13,21 +14,22 @@ interface IMqProducer {
     /**
      * 注册主题
      * @param topic 主题
-     * @return
+     * @return false表示没有broker可分配
      */
-    fun registerTopic(topic: String)
+    fun registerTopic(topic: String): Boolean
 
     /**
      * 注销topic
      *
      * @param topic
-     * @return
+     * @return false表示topic根本就没有分配过
      */
-    fun unregisterTopic(topic: String)
+    fun unregisterTopic(topic: String): Boolean
 
     /**
      * 发送消息
      * @param msg 消息
+     * @return broker生成的消息id
      */
-    fun send(msg: Message)
+    fun send(msg: Message): CompletableFuture<Long>
 }