+++ jksoa-mq/jksoa-mq-registry/src/main/kotlin/net/jkcode/jksoa/mq/registry/IMqDiscoveryListener.kt
@@ -1,7 +1,7 @@
 package net.jkcode.jksoa.mq.registry
 
 /**
- * 处理topic分配变化
+ * topic分配变化的监听器
  *
  * @author shijianhang
  * @create 2017-12-13 下午10:38