+++ jksoa-mq/jksoa-mq-registry/src/main/kotlin/net/jkcode/jksoa/mq/registry/zk/ZkMqRegistry.kt
@@ -1,6 +1,8 @@
 package net.jkcode.jksoa.mq.registry.zk
 
 import net.jkcode.jksoa.common.Url
+import net.jkcode.jksoa.mq.common.TopicRegex
+import net.jkcode.jksoa.mq.common.mqRegisterLogger
 import net.jkcode.jksoa.mq.registry.*
 import net.jkcode.jksoa.registry.IRegistry
 import net.jkcode.jksoa.registry.zk.ZkRegistry
@@ -50,9 +52,7 @@ object ZkMqRegistry: ZkMqDiscovery(), IMqRegistry {
         assigner.assignTopic(topic)
 
         // 写topic分配
-        val json = assignment.toJson()
-        zkClient.writeData(topic2brokerPath, json)
-        mqLogger.info("给topic[{}]分配broker: {}", topic, json)
+        persist(assignment)
         return true
     }
 
@@ -75,7 +75,7 @@ object ZkMqRegistry: ZkMqDiscovery(), IMqRegistry {
         assignment.remove(topic)
 
         // 写topic分配
-        zkClient.writeData(topic2brokerPath, assignment.toJson())
+        persist(assignment)
         return true
     }
 
@@ -107,7 +107,23 @@ object ZkMqRegistry: ZkMqDiscovery(), IMqRegistry {
             assigner.assignTopic(topic)
 
         // 写topic分配
-        zkClient.writeData(topic2brokerPath, assignment.toJson())
+        persist(assignment)
         return true
     }
+
+    /**
+     * 写topic分配
+     * @param assignment
+     */
+    private fun persist(assignment: TopicAssignment) {
+        // 写topic分配
+        val json = assignment.toJson()
+        zkClient.writeData(topic2brokerPath, json)
+        mqRegisterLogger.info("topic分配结束: {}", json)
+
+        // 主动触发本地监听器
+        for (l in discoveryListeners())
+            l.handleTopic2BrokerChange(assignment)
+    }
+
 }