+++ jksoa-registry/src/main/kotlin/net/jkcode/jksoa/registry/zk/ZkDiscovery.kt
@@ -42,6 +42,15 @@ open class ZkDiscovery: IDiscovery {
     protected val dataListeners = ConcurrentHashMap<String, ConcurrentHashMap<IDiscoveryListener, List<ZkDataListener>>>()
 
     /**
+     * 获得监听器
+     * @param serviceId
+     * @return
+     */
+    public override fun discoveryListeners(serviceId: String): Collection<IDiscoveryListener> {
+        return dataListeners[serviceId]?.keys ?: emptyList()
+    }
+
+    /**
      * 监听服务变化
      *
      * @param serviceId 服务标识